requirements.txt
temporalio==1.6.0
python-cowsay==1.2.1
fastapi==0.104.1
uvicorn[standard]==0.24.0